Still fixing WSL2 resolv.conf issues. For idempotency resolv.conf will be fixed only if it is a symlink when running ubuntu_deployer-setup.sh

+++ ubuntu_deployer/ubuntu_deployer-setup.sh
@@ -84,20 +84,29 @@ fi
 ############################
 # DNS workaround for resolution with SAS VPN and WSL issues
 ############################
-echo "[INFO] Running DNS fix for SAS VPN and WSL issues"
-sudo unlink /etc/resolv.conf # this will unlink the default wsl2 resolv.conf
-sudo cp /run/resolvconf/resolv.conf /etc/resolv.conf
-sudo perl -p -i -e 's/nameserver 172\.26\.40\.178/nameserver 8.8.8.8\nnameserver 172.26.40.178/' /etc/resolv.conf
+if [ -L  /etc/resolv.conf ]
+    then
+    echo "[INFO] Running DNS fix for SAS VPN and WSL issues"
+    sudo unlink /etc/resolv.conf # this will unlink the default wsl2 resolv.conf
+    # sudo cp /run/resolvconf/resolv.conf /etc/resolv.conf
+    # sudo perl -p -i -e 's/nameserver 172\.26\.40\.12/nameserver 172.26.40.178\nnameserver 8.8.8.8/' /etc/resolv.conf
+    sudo cat <<EOF | sudo tee /etc/resolv.conf
+nameserver 172.26.40.178
+nameserver 172.26.40.12
+nameserver 8.8.8.8
+search sas.com
+EOF
 
-cat <<EOF | sudo tee /etc/wsl.conf
+    cat <<EOF | sudo tee /etc/wsl.conf
 [network]
 generateResolvConf = false
 EOF
 
-# Make resolv.conf immutable so WSL doesn't clobber it on startup
-echo "[INFO] Make resolv.conf immutable to prevent clobbering by WSL on startup"
-sudo chattr -f +i /etc/resolv.conf;
+    # Make resolv.conf immutable so WSL doesn't clobber it on startup
+    echo "[INFO] Make resolv.conf immutable to prevent clobbering by WSL on startup"
+    sudo chattr -f +i /etc/resolv.conf
 
+fi
 
 echo "[INFO] Installing pre-requisite OS packages."
 # Install pre-requisite OS packages
